+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -------------------------
 - Enh #163: Attachments for HTML pages
 - Enh: HumHub min. version increased to 1.8
+- Enh: Improved page edit form with beginCollapsibleFields
 
 
 1.3.1 (February 23, 2021)